Restructure the code based on go lang conventional
